--- scripts/transfer_ownership.js
@@ -1,16 +1,0 @@
-// scripts/transfer_ownership.js
-async function main() {
-    const gnosisSafe = '0x5288d36112fe21be1a24b236be887C90c3AE7090';
-   
-    console.log("Transferring ownership of ProxyAdmin...");
-    // The owner of the ProxyAdmin can upgrade our contracts
-    await upgrades.admin.transferProxyAdminOwnership(gnosisSafe);
-    console.log("Transferred ownership of ProxyAdmin to:", gnosisSafe);
-  }
-   
-  main()
-    .then(() => process.exit(0))
-    .catch(error => {
-      console.error(error);
-      process.exit(1);
-    });
